User space threads (aka 'green threads') support for `linux-thread-db`.
This patch extends `linux-thread-db` to support user space threads
running in addition to LWP kernel threads. This is opted into by
the user supplying a custom `libthread_db.so.1` using the GDB
command `set libthread-db-search-path`. If that custom
`libthread_db.so.1` reports threads of type `TD_THR_SYSTEM`
(NPTL only ever reports type `TD_THR_USER`), it turns on the user
space threading support.

If that support is enabled, `linux-thread-db` will now additionally
ask `libthread_db.so.1` about the state of the user space threads
only using the filter `state = TD_THR_RUN`. It will add and remove
those from the GDB thread list as needed. `info threads` will now
show which type of thread each is, either system or user.

You can use the GDB command `thread` to examine the current state
of a suspended user space thread, including backtracing it. If
a user space thread is currently running on a LWP, it shows as
'running' in the thread list.

I have supplied an example implementation of how userspace would
communicate with a custom `libthread_db.so.1` in the file:

gdb/linux-thread-db-user-threads.h

This aims to be race free with respect to concurrent reads of
the inferior by GDB. Using that example implementation, a test
exercises an inferior which creates a user space thread.

The test code requires platform specific code, so I gated it to
x64 and AArch64 only, as those are the only we care about. I
don't actually have an AArch64 system to test this upon, so
I would entirely understand if feedback requests removing that
architecture.